feat: delete resources folder
