+++ import-service/tsconfig.paths.json
@@ -4,7 +4,6 @@
     "paths": {
       "@functions/*": ["src/functions/*"],
       "@libs/*": ["src/libs/*"],
-      "@resources/*": ["src/resources/*"]
     }
   }
 }
